Changed to use gormigrate
